requirements.txt
streamlit==1.28.2
openai==1.3.6
pandasai==1.5.4
pandas==1.5.3
python-dotenv==1.0.0
google-generativeai==0.3.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
seaborn==0.12.2
plotly==5.18.0
